fix(fe): componentize contest tabs
